web/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.1.3
numpy>=1.24.3

